convert search to api.js and fix a bug in photos dataIn (need to require logger)
